+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Add the `reduce()` function : Reduces an array by removing values based on conditions or using compress/clean.
   - Add the `reorder()` function : Reorders an associative array by placing specified keys first, optionally sorting the rest.
   - Add the `merge()` function with the `MergeOption` and `NullsOption` helpers.
+  - Add the `prepare()` function.
 - **Strings**
   - Add the `chunk()` function : Splits a string into groups of length, separated by a separator.
   - Add the `replace()` function : Replaces all occurrences of a substring within a string, with optional Unicode (grapheme-safe) and normalization support.